# Conflicts:
#	ERD/table_ctt.erd
#	src/main/java/com/lec/spring/controller/mypage/MyPageController.java

+++ src/main/java/com/lec/spring/domain/order/Order.java
@@ -14,8 +14,8 @@ import java.time.LocalDateTime;
 @Builder
 public class Order {
 
-    private int id;
-    private int user_id; // 유저 아이디
+    private Long id;
+    private Long user_id; // 유저 아이디
     private int number; // 주문 묶는 번호
     private LocalDateTime regdate; // 주문 일시
 }
